fix(math): emit latex wrapped in brackets
The math plugin was emitting raw latex, but needs to wrap it in brackets like so: '\(latex\)'.
Changed the attribute name to 'data-latex' instead of 'data-mathjax'.

Note: there is an outstanding issue that some characters are escaped in the latex like: `&lt;`, but it doesn't affect Katex rendering so going to proceed with this release.