feat: new charting/graphing
* new @pie-lib/charting - new charting components to more closely match
our needs.
* new @pie-lib/graphing - new grahing components
* new @pie-lib/plot - a library with shared modules for the above.

Note that both libs are incomplete in terms of requirements, but have
the building blocks in place for more features to be added easily.

See the demo site on examples of using them.

BREAKING CHANGE: Charting is a complete re-write and so any downstream
libs will need to be updated to work with them.